graphics and game manager added
fleshed out graphics manager. fixed manager to not be a singleton so you
can have more than one at a time, in case you want to flip between
scenes dynamically and not have to reload graphics each time.

Added a game manager to manage the game's states and rendering.

Added a game settings file to manage global settings for the game.