transitioned game manager to be main interface
GameManager is now the primary interface for rendering and updating. It
now calculates deltaTime since the last update and keeps consistent
track of time. It also handles all rendering and SpriteManager
interfacing in preparation to transition over to a state-based system.

Also removed geometric background drawing from the sprite manager as it
was causing the redraw to flicker.